+++ ClinicaMVC/src/main/java/dh/backend/maxisoriano/ClinicaMVC/service/impl/PacienteService.java
@@ -3,8 +3,11 @@ package dh.backend.maxisoriano.ClinicaMVC.service.impl;
 
 
 import dh.backend.maxisoriano.ClinicaMVC.entity.Paciente;
+import dh.backend.maxisoriano.ClinicaMVC.exception.ResourceNotFoundException;
 import dh.backend.maxisoriano.ClinicaMVC.repository.IPacienteRepository;
 import dh.backend.maxisoriano.ClinicaMVC.service.IPacienteService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,7 +22,6 @@ public class PacienteService implements IPacienteService {
     }
 
     public Paciente registrarPaciente(Paciente paciente){
-
         return this.pacienteRepository.save(paciente);
     }
 
@@ -35,7 +37,22 @@ public class PacienteService implements IPacienteService {
         this.pacienteRepository.save(paciente);
     }
     @Override
-    public void eliminarPaciente(Integer id) {
-        this.pacienteRepository.deleteById(id);
+    public void eliminarPaciente(Integer id) throws ResourceNotFoundException {
+        Optional<Paciente> pacienteOptional = this.buscarPorId(id);
+        if (pacienteOptional.isPresent()) {
+            this.pacienteRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("{\"message\": \"Paciente no eliminado\"}");
+        }
+    }
+
+    @Override
+    public List<Paciente> buscarPorDni(String dni) {
+        return this.pacienteRepository.buscarPorDni(dni);
+    }
+
+    @Override
+    public List<Paciente> buscarPorProvincia(String provincia) {
+        return this.pacienteRepository.buscarPorProvinciaLike(provincia);
     }
 }
